requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.51.0
langgraph==0.2.0
pydantic==2.8.2
python-dotenv==1.0.1